py/Snippy/requirements.txt
PyQt5==5.15.9
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pyautogui==0.9.54
mss==9.0.1
pyinstaller==6.3.0
